Simplify structure declarations.
Many structures in the code are defined in the form:

    typedef struct A {...} AA;

In C++ this is redundant, creating both the names A and AA that can be
used interchangeably.  Simplify these to the form:

    struct AA {...};

These changes were inspired by the clang-tidy "modernize use using"
checker which looks to replace the use of the "typedef" keyword with
the newer "using" keyword.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-using.html