Tidy: Replace "typedef" with "using" in header files.
The clang-tidy modernize "use using" checker pointed out a number of
places where the older style of "typedef a b;" should be replaced with
the modern construct of "using b = a;", but it didn't point out many
instances in header files, only in cpp files.  Make these replacements
by hand in header files.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-using.html