Extend instance locking in HDHomeRun stream handler
Extend the instance locking already present in HDHRStreamHandler
functions TunerGet and TunerSet to GetTunerStatus.
This fixes a segfault in libhdhomerun, reported in ticket #13519,
that appeared when doing a UK DVB-T/T2 scan with offset frequencies.
The segfault happened because GetTunerStatus was called from the
HDHRSignalMonitor in a different thread than the TunerGet/TunerSet functions.
Extend the instance locking also to the HDHRStreamHandler::run function
where the libhdhomerun functions are called. This should fix the segfault
in the mythbackend startup as reported in ticket #9027.

Refs #9027

Fixes #13519