Tidy: Fix const-ness of variables and functions.
The const_cast conversion should never be used.  Casting a const
variable to a non-const variable and then operating on it is undefined
behavior.

One surprising revelation is that calling MythTimer::elapsed to see if
the timer has expired can restart the timer.  Changing this function
take a non-const timer causes ripples through the code.

Another changes that causes a few ripples is that
ProgramInfo::SetPathname modifies the ProgramInfo object.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-const-cast.html