Tidy: Fix constness of MPEGStreamData::ReturnCachedTable.
The const_cast conversion should never be used.  Casting a const
variable to a non-const variable and then operating on it is undefined
behavior.

This function casts away the constness of a parameter so that it can
call MPEGStreamData::DeleteCachedTable.  Fix this by modifying
DeleteCachedTable to take a const pointer.  That function may delete
the pointer, but that's apparently OK according to the standard.  The
"const" keyword only specifies that you cannot modify the object, it
implies nothing about deleting the object.  This change ripples to all
the functions that overload DeleteCachedTable.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-const-cast.html