Tidy: Don't use floating point variables as loop counter.
The clang-tidy "cert-flp30-c" checker pointed out two locations where
a floating point variable was used as a loop counter.  Because of the
limitations of floating point variables, it is considered bad practice
and non-portable to use them as loop counters.  These two instances
are trivial enough that it shouldn't be a problem, but its easy to
correct them to use an integer loop counter.

https://clang.llvm.org/extra/clang-tidy/checks/cert-flp30-c.html