Tidy: Reduce excess padding in the MythMainWindowPrivate structure.
The clang-tidy "excessive padding" checker pointed out that the
MythMainWindowPrivate structure contained 38 bytes of padding, where
an optimal ordering of the member variables would produce only 6 bytes
of padding.  Groups a number of the bool variables together to reduce
the padding to the point where the checker doesn't complain anymore.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-optin.performance.Padding.html