Tidy: Use the auto type specifier instead of duplicating types. (libmythtv)
The clang-tidy "use auto" modernization checker pointed out a ton of
places where a type specifier was duplicated, e.g. "Foo *var = new
Foo();", where the code can be simplified by using the new "auto"
keyword.  The previous example will be updated to read "auto *var =
new Foo();".  This example looks trivial, but when "Foo" is a 25
character QMap or something even longer, it makes the code easier to
read.  All replacements made by clang-tidy with some manual
reformatting.

There are a couple of manual replacements of begin()/end() with
cbegin()/cend() to ensure that iterators are still const when using
the "auto" keyword.

There is also one tweak for a place where a pointer was malloced with
"sizeof(*ptr)" as the number of bytes of data.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-auto.html