Add cbegin/cend functions to AutoDeleteQueue and use in programinfo.cpp.
The clang-tidy "use auto" modernization checker pointed out one places
in programinfo.cpp where a type specifier was duplicated and attempted
to replace it with the "auto" keyword.  The automatic replacement
would have lost the constness of the variable, so add create
cbegin/cend functions to the AutoDeleteQueue template and use those
functions with the auto keyword.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-auto.html