Rename the association_list const find function, and use with "auto".
The clang-tidy "use auto" modernization checker pointed out a couple
of places in dbaccess.cpp where a type specifier was duplicated and
attempted to replace it with the "auto" keyword.  These automatic
replacements would have lost the constness of the variable, so rename
the const versions of the "find" function to"cfind" and use those
functions with the auto keyword.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-auto.html