Fix null pointer dereferences in xine_demux_sputext.cpp.
The clang-tidy "null pointer dereference" checker pointed out a couple
of places in this file where a null pointers can be dereferenced.
This is caused by callers of the sub_readtext function checking the
wrong location for error returns.  The function returning a value of
-1 on error, but callers are checking the value of the seconds
argument for a -1 error return.  Fix the callers to check the return
value.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-core.NullDereference.html