tidy: Fix clang-tidy warning about setting a variable and then never using it.
The clang-tidy "dead stores" checker pointed out a number of places
where a valueis stored into a variable, and then that variable isn't
accessed before the end of the funcion, or before it is overwritten
with a new value.  Remove most of these dead stores, and comment out
the others.

https://clang.llvm.org/extra/clang-tidy/checks/readability-inconsistent-declaration-parameter-name.html