Transport Editor and HDHOMERUN devices
A HDHOMERUN device can be DVB-C, DVB-T/T2, ATSC or a combination of those.
The Transport Editor checks, before starting the editor,
if all the capture cards that are connected to that video source
do have the same type.
In this check all HDHOMERUN devices were seen as ATSC devices.
This creates a conflict when a HDHOMERUN device is connected to
a videosource together with e.g. a DVB-C device.
The test now assumes that a HDHOMERUN device can be anything.