tidy: Always allocate the max size array for volumes
Clang-tidy complains that the AudioOutputJACK::DeinterleaveAudio
function allocates a variable length array of negative size using this
statement:

    float volumes[m_channels];

Presumably the value of m_channels will have been properly set by the
time this function is called, but its default value is -1 which makes
the clang-tidy complaint reasonable.  Given that the max possible
number of channels for JACK is 8, just use this value when allocating
the array.  It may be too large in some instances, but it will never
be too small.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-core.VLASize.html