MythDisplayX11: Use accurate XRandR refresh rates
- use the more modern XRandR API to interrogate actual modelines
available and hence calculate refresh rates with full floating point
accuracy
- enables the removal of libxnvtrl at some point
- working on multiple testes linux platforms
- there is a fairly ugly workaround to ensure the main window is not
clipped when we have moved to a higher resolution mode - if necessary,
we hide and then show the main window. I have no idea why this is
necessary and can't work out why it is needed - all the loggging/
debugging in the world doesn't indicate why sometimes the window is
clipped.
- also moves a little state management out of MythUIHelper and
MythVideoOutput - which is a good thing.

Refs #13100