Show name of video source in Transport Editor
In the mythtv-setup Transport Editor, the currently selected video
source was shown in an selection box in which selecting another
value was made impossible by setting the selection box to disabled.
In a disabled entry the label is shown greyed out which is not
intended here; it is the intention that the value should not be modified.
Also, a disabled entry cannot be navigated to with the arrow keys
and this means that the associated help text will also not be shown.

The name of the currently selected video source is now shown
in a read-only field.