Improve and fix decoder callbacks into the UI thread
- I could write a book on this one
- suffice it to say that using the MythMainWindow instance to process
decoder callbacks was a convenient sledgehammer when the more involved
use of MythPlayer itself gives the precision we need.
- fixes sundry decoder lockups caused by calling qProcessEvents when
trying to ensure the decoder callbacks are processed - qProcessEvents is
the aforementioned sledgehammer...
- forces a couple of capability checks to become UI thread only (which
is fine) because there is no player to access.
- otherwise we always need acccess to a MythPlayer instance - which for
the decode side is available from the decoder but must be added to the
interop side