Increase default tuning timeout values
For DVB-T/T2/C devices, increase the default signal timeout values from 1000ms
to 3000ms and the default tuning timeout values from 3000ms to 6000ms.
For ATSC devices, increase the default signal timeout value from 500ms to
2000ms and the default tuning timeout value from 2000ms to 4000ms.
This does increase the chance of getting good signal especially on DVB-T/T2.
This does increase the channel scanning time when there is no signal
on the frequency that is tuned to, which happens when doing a "Full Scan"
in which a list of frequencies is tried just to see if there
is something present there.
Note that this are the default values for new capture cards.
The timeout values of existing capture cards is not changed.
It is always possible to change the timeout values in the Capture Card settings page.

Refs #13472