tidy: Pass constructor arguments by value and use std::move.
The clang-tidy "pass by value" modernization check pointed out a
number of places where constructors are declared to take a const
object, and the the only use of that object is to copy it to a newly
initialize local variable.  Passing a non-const and using the
std::move function allows the compiler to choose the best way to
construct the local copy of the value.  Changes made by the clang-tidy
program and cleaned up by hand.

https://clang.llvm.org/extra/clang-tidy/checks/modernize-pass-by-value.html