tidy: Convert member functions to static if they don't use "this".
The clang-tidy readability "convert to static" checker pointed out a
large number of places where a member function doesn't reference the
current object, and thus doesn't need to be provided with a "this"
variable that points to an object.  These function should all be
declared as static, instead of as const.  Declaring a function as
"const" is a statement that it won't modify the object, but it still
possible for the function to somehow affect it.  A function declared
"static" doesn't receive a pointer to the object and so can't possibly
modify it.  Changes made by the clang-tidy program.

https://clang.llvm.https://clang.llvm.org/extra/clang-tidy/checks/readability-convert-member-functions-to-static.html