tidy: Clean up control flow jumps followed by else statements.
If the last statement in an "if" block unconditional changes the
control flow, then there is no need to follow that statement with an
else block.  Omitting the "else" statement and the indentation of the
subsequent block makes the code easier to follow.  These changes can
be easier to read using the "-b" flag to diff.

https://clang.llvm.org/extra/clang-tidy/checks/readability-else-after-return.html