tidy: Mark some additional constructors as "explicit". (services)
The clang-tidy "explicit constructor" checker pointed out a couple of
places where constructors should be marked as explicit.  It also
falsely pointed out a couple of class "operator" methods that it
thought should be explicit, but marking these as explicit causes the
compilation to fail in unexpected ways, so mark them to be ignored
instead.

https://clang.llvm.org/extra/clang-tidy/checks/google-explicit-constructor.html