Change "Wrong PMT" messages from LOG_ERROR to LOG_DEBUG
The "Wrong PMT" messages from DTVSignalMonitor::HandlePMT appear
always when changing channels in Live TV.
What happens is that all programs in the PAT of that multiplex are
tried in sequence to find the correct channel. This is done only
in Live TV, not in recordings.
As this message gives information about the flow of the
program execution and does not indicate an error the message
is now categorized as LOG_DEBUG and not as LOG_ERROR.
Also, the error class is now set to CHANNEL instead of GENERAL.