Update qmake .pro file modernization for various non-app directories
Change qmake TEMPLATE to aux (from app) for directories that do not
build anything (and remove app template artifacts).

The aux template was added in QT5.  Per the docs the aux template
"Creates a Makefile for not building anything. Use this if no compiler
needs to be invoked to create the target" which is exactly what we
want for these directories.

Fixes #13542.

Signed-off-by: David Hampton <mythtv@love2code.net>