Fix clang-tidy warning about dereferencing an undefined pointer.
The clang-tidy "Null Deference" checker is errantly complaining that
the Jack code can dereference an undefined pointer.  This is happening
because tidy appears to be using the default value of -1 for
m_channels when looking at the code that allocates buffers, and then
using a real value like 2 for m_channels when looking at the code that
writes into the buffers.  Zeroing out the entire array of allocated
buffer pointers eliminates this warning message.