Increase default tuning timeout values for HDHomeRun devices
The signal timeout value is changed from 1000 to 3000 ms.
The tuning timeout value is changed from 3000 to 6000 ms.
This was already done for conventional DVB-T/T2 and DVB-C tuners.
This does increase the chance of getting good signal especially on DVB-T/T2.
This does increase the channel scanning time only when there is no signal
on the frequency that is tuned to, which happens when doing a "Full Scan"
in which a list of frequencies is tried just to see if there
is something present there.
Note that this are the default values for new capture cards.
The timeout values of existing capture cards is not changed.
It is always possible to change the timeout values in the Capture Card settings page.