tidy: Fix unnamed arguments on functions.
The clang-tidy "constant parameter in declaration" readability check
pointed out a number of places where a function parameter was declared
constant, but that declaration is irrelevant to the signature of the
function.  These mostly appear to be integer parameters being declared
as constant, which is redundant since integers are always passed by
value and not by reference.  Remove these unnecessary "const" keywords
for readability.

https://clang.llvm.org/extra/clang-tidy/checks/readability-avoid-const-params-in-decls.html