tidy: Avoid warning of a variable being declared in a header file.
The clang-tidy "misc definition in header" checker looks for variables
that are defined in header files, which can lead to multiple
definitions of the variable.  This check mistakenly flags the variable
gBackendURI as a problem, even though it is const.  Rename this
constant to begin with the letter 'k' since that is what is used
throughout the program for constants.

https://clang.llvm.org/extra/clang-tidy/checks/misc-definitions-in-headers.html