tidy: Avoid warning of a variable being declared in a header file #2.
The clang-tidy "misc definition in header" checker looks for variables
that are defined in header files, which can lead to multiple
definitions of the variable.  The mpeg freesat tables are declared in
a header file (and are the only thing in the header file), and this
file is only included once so there is no problem.  A quick solution
to this errant warning is simply to rename the file from
freesat_tables.h to freesat_tables.cpp.

https://clang.llvm.org/extra/clang-tidy/checks/misc-definitions-in-headers.html