OpenTV logical channel numbers
Support for OpenTV logical channel numbers as used
by Sky (US/IT/NZ/AU), Dishnet(US) and others.
The Bouquet ID and the Region ID, needed to get the correct
tables, can be defined in the video source  similar to how it
is done for the Freesat and Sky UK channel numbers.
Scanning is activated by a new "Scan OpenTV" channel scanning option.
As a next step, when the applicable network ID's and private descriptor
identifiers are known, the identification of a stream with OpenTV
channel numbers can be done automatically.

Refs #13472
Fixes #13547

Signed-off-by: Klaas de Waal <klaas@kldo.nl>