Remove debug output of OpenTVChannelListDescriptor 0xB1
In mythtv-setup, remove the "toString" debug output of
descriptor 0xB1 as an OpenTVChannelListDescriptor.
When scanning the Hotbird 13E0 satellite this causes a segfault.
The debug output for the equivalent BSkyBLCNDescriptor, also 0xB1,
as used on the Astra-2 28E2 satellite, is still present;
this one is only used when preceded by the Private Data
Specifier for BSkyB, 0x00000002, as defined in
https://www.dvbservices.com/identifiers/private_data_spec_id.