MythVideoOutput: Rework aspect ratio code (with db schema update)
- N.B. The behaviour of 'Screen aspect ratio' setting is changed
(formerly 'Monitor aspect ratio' in 0.30 and 'Virtual monitor aspect
ratio' in early pre-0.31 versions). This now defaults to a value of
'auto' and other specific overrides are available when automatic
detection does not work.
- As this setting would probably have had a spurious default previously
(and would have been generally unused), a database schema update resets
this to 0.0 (which is auto).
- otherwise this change consolidates the display aspect ratio code into
MythDisplay - where the setting precedence etc is clearly documented.
- all of the old, unnecessary calculation of physical sizes etc in
MythVideoOutput is gone - all it needs is the correct aspect ratio.

- also fixes a regression in the screen switching code