tidy: operator++ and operator-- should return const objects.
The clang-tidy "cert-dcl21-cpp" checker pointed out two places where
operator++ and operator-- functions don't return const objects.  Since
the return value is supposed to be the value of the object *before*
the increment or decrement, any change to the returned value is a
change to a temporary object and will be lost.  Mark these objects as
const to prevent and modification.

https://clang.llvm.org/extra/clang-tidy/checks/cert-dcl21-cpp.html