tidy: Mark "auto" variable as "const" or pointer.
The new clang-tidy-11 "qualified auto" checker pointed out many places
where the deduced variable was a pointer to an object, so "auto foo"
should be replaced with "auto *foo" to make is obvious that the
variable is a pointer.  There were also a few instances where the the
deduced variable was a const pointer to an object, so in these cases
"auto foo" should be replaced with "const auto *foo".  These changes
don't have any affect on the resulting executable; they only affect
the readability of the code.  All changes made by clang-tidy.

https://clang.llvm.org/extra/clang-tidy/checks/readability-qualified-auto.html