tidy: Fix uninitialized variable warning in vbi.cpp.
The clang-tidy "uninitialized variable" checker pointed out one place
where the vbi code was clearing a bit in an uninitialized variable.
Converting this code to use the C++ new/delete operators fixes this
problem, and also furthers another goal which it to replaces all calls
to malloc/free in C++ code with the new/delete operators.

This change also adds call to delete the vbi_client object when it is
no longer needed. This appeared to be a memory leak in the original
code.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-core.uninitialized.Assign.html