tidy: Fix uninitialized variable/branch warning in xine_demux_sputext.
The clang-tidy "uninitialized variable" checker pointed out where the
code appears to be using an uninitialized variable.  That doesn't
appear possible from my reading of the code.  The function in question
strips whitespace form the beginning and end of a string.  It appears
that all callers terminate the string they are building before calling
this function. In the degenerate case, the string is empty so the only
character is the NUL terminator character, so the copy loop is never
called and thus no access to an uninitialized variable copy[1].  With
any non-empty string, copy[1] is always guaranteed to be another
character or the NUL terminator, and therefore is always initialized.
Tell clang-tidy to ignore this line of code in the future.

The clang-tidy "uninitialized branch" checker pointed out where the
code was branching on a value in a newly malloced data structure.
Given that the standard doesn't require malloced memory to be set to
zero, even though it often seems to be, this value is undefined.  Fix
the warning by guaranteeing that the memory is initialized with zeros.
This memory can be reallocated, so there are two calls to malloc.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-core.uninitialized.Assign.html
https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-core.uninitialized.Branch.html