MythFrontend: Convert ExitPrompter to use MythPower
- most, if not all, users should see no change to existing behaviour
- if a command is set for reboot or shutdown, these are still used
- the old dbus code used a mix of logind, HAL, ConsoleKit and UPower
interfaces
- consolekit is now no longer installed on most systems and is
unmaintained and considered unsafe
- HAL is deprecated
- UPower no longer supports suspend or hibernate
- and logind is now the interface of choice
- MythPower also has an OSX implementation offering similar functionality
- add a new option for Suspend
- extend the confirmation dialog so that it is always used when a
shutdown/suspend/reboot request is initiated
- there is a default 3 second delay between the request to MythPower and
its implementation. When MythPower is added by default to
MythCoreContext, this will allow other MythTV components to request that
the delay is extended so that shutdowns etc can be handled gracefully
(e.g. allow SSDP Bye Bye notifications to be sent twice over a 5 second
interval).
- the use of logind scheduled requests also allows other system
components to handle shutdown etc gracefully if need be.
- it is theoretically possible to add another dialog that would offer a
countdown and ability to cancel the shutdown/reboot/suspend - but most
window managers appear to start shutting down display etc as soon as the
request is made - and hence the extra dialog is not isible/useful