Add a new hack to make sure an old, live TV hack is performed.
Sometimes, m_framesPlayed > frameCount in MythPlayer.  When that
happens in live TV or watching while recording, the code to
temporarily pause playback to prevent stuttering never gets triggered.
This change detects that case so the pause occurs.  Without it, the
resulting stuttering can go on indefinitely.

Ultimately, we still need to find out why the frame counts are off.