configure: Link to either OpenGL or OpenGL ES - not both
- linking to both libraries causes unexpected behaviour on some legacy
drivers - notably nvidia (and maybe some Intel mesa) - leading to either
a hard crash at startup or falling back to llvm pipe (software
rendering)
- it also appears to impact behaviour of EGL DMABUF support on the Pi4 -
which needs an additional fix following this change
- we default to OpenGL and check whether Qt was compiled for GLES2 - and
if so only search for GLES2.
- in case Qt support is misdetected an additional configure option is
added to force OpenGL ES
- note however that this is ONLY intended as a workaround for
misdetection - Qt will still load the version of OpenGL it was compiled
for. Linking to GLES is not required to actually use GLES on desktop
platforms (this can be enabled with 'MYTHTV_NO_EGL=1
MYTHTV_OPENGL_ES=1')
- refs #13538