Fix parameter confusion to MythImage::MakeGradient.
New optional parameters were added to this function in 2008, and not
all callers got updated properly.  Implicit type conversions between
"int" and "bool" meant that the two new arguments were not
distinguishable, so when the first was omitted the second was
implicitly converted and used as the first.

This commit adds the missing parameter, and converts the code to use
"enum classes" to prevent this problem from happening again.