tidy: Convert variable int->bool in scheduler.
The clang-tidy "implicit boolean conversion" check pointed out a
number of places where a set of boolean comparisons were performed,
and then the result was implicitly converted to an integer.  Change
the variable types to "bool" so that the conversions don't have to be
made.  No other changes to the scheduler code.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html