tidy: Fix some calls objects passed by throw and catch.
The clang-tidy "throw by value, catch by reference" checker pointed
out a one place (in a macro) where a QString was caught as an object,
instead of as a reference to an object.  Change this to a reference,
as that only requires assigning a pointer instead of copying the
entire object.

The checker also pointed out several places where named objects were
thrown, instead of anonymous temporary values.  Its likely that an
anonymous object allows the compiler to simply pass the object
directly to the catcher, instead of having to create its own anonymous
copy.  Change these calls to throw anonymous objects.

https://clang.llvm.org/extra/clang-tidy/checks/misc-throw-by-value-catch-by-reference.html