tidy: Eliminate some "narrowing conversion" warnings.
The clang-tidy "narrowing conversion" checker pointed a number of
places where implicit type conversions reduce the size or precision of
the variable.  For example, an unsigned integer being cast to an
integer is only valid for half of the possible range of values of the
unsigned integer.

This commit changes the type of a number of variable, so that all
assignments and comparisons happen without an implicit type
conversion.  It also adds some explicit casts, where the variable type
can't be changed.  I.E. The "uint" return value of one library call
that is passed to a second library that expects an "int".

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-narrowing-conversions.html