tidy: Remove many __XXX_SOURCE identifiers
The new clang-tidy-11 "reserved identifier" checker looks for code
that violates the C/C+ standards for identifier naming.  The C++
standard reserves 1) all names that start with an underscore followed
by an uppercase letter, and 2) any name that contains a double
underscore.  Use of these names results in undefined behavior.

There are a number of identifiers of the form _XXX_SOURCE that specify
support for additional feature that became available in specific
versions of the C/C++ standards.  These are unnecessary as the
settings.pro file provides _GNU_SOURCE, and when the
/usr/include/features.h file sees _GNU_SOURCE defined, it will define
the following:

   _ISOC95_SOURCE, _ISOC99_SOURCE, _ISOC11_SOURCE, _POSIX_SOURCE,
   _POSIX_C_SOURCE, _XOPEN_SOURCE, _XOPEN_SOURCE_EXTENDED,
   _LARGEFILE64_SOURCE, _DEFAULT_SOURCE, _ATFILE_SOURCE,

Remove these identifiers from the MythTV sources.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-reserved-identifier.html