tidy: Remove excessive padding in the DecoderBase class.
The clang-tidy "padding" checker looks for classes and structures that
are excessively padded due to misalignment of member sizes and their
offsets within a structure.  According to tidy, the DecoderBase class
has 40 bytes of padding, where a complete reordering of the class
members would have no wasted space.  Do some local reordering to lower
the amount of wasted space below the checker's default threshold of 24
bytes.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-optin.performance.Padding.html