tidy: Remove excessive padding in the MythPlayer class.
The clang-tidy "padding" checker looks for classes and structures that
are excessively padded due to misalignment of member sizes and their
offsets within a structure.  According to tidy, the MythPlayer class
has 113 bytes of padding, where a complete reordering of the class
members would reduce this to only one byte of wasted space.  Do some
local reordering to lower the amount of wasted space from 113 to 73
bytes, which is still above the checker's default threshold of 24
bytes.  Disable the check for now.

https://clang.llvm.org/extra/clang-tidy/checks/clang-analyzer-optin.performance.Padding.html