tidy: Clean up most static downcast warnings from qApp.
The clang-tidy "static downcast" checker warns on every use of the
"qApp" macro, because that macro contains a static cast from
QCoreApplication to the grandchild type QApplication.  Many of the
places that use this macro then immediately call a static function on
the grandparent QCoreApplication class or the parent QGuiApplication
class, which should also get flagged by the "static accessed though
instance" checker.  Replace this combo "qApp->fn" call with a direct
call to the static Q*Application function.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-static-cast-downcast.html