One-part channel number for ATSC cable
Implement the one-part channel number for ATSC cable as defined
in ATSC standard A65/2013  7 August 2013  page 35.
In this case the ATSC major number is 1008 (0x3F0) or higher
and all channels in all multiplexes have this major number.
Change the name of the MPTS full-transport stream channel
to use the freqid instead of the major number when
the one-part channel number is used.