tidy: Clean up remaining static downcast warnings from qApp.
The clang-tidy "static downcast" checker warns on every use of the
"qApp" macro, because that macro contains a static cast from
QCoreApplication to the grandchild type QApplication.  Replace the
remaining uses of the qApp/qGuiApp macros with explicit conversions
via the qobject_cast() function, and add tests for conversion
failures.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-static-cast-downcast.html