mythmusic: include discnumber in album sort order
Album sort did not handle disc number correctly, e.g. for a two disc album it
would sort disc.track-wise as 1.1, 2.1, 1.2, 2.2, etc rather than 1.1, 1.2,
..., 2.1, 2.2.

Incorporate the disc number into the album order to correct this. I believe a
single CD can have at most 99 tracks so use a multiplier of 100. This means a
10 disc box set would clash with album map order offset of 1000, so bump that
to 10000 to leave space for up to a 100 disc box set.