musicmetadata: ensure compilation artist id is always set
I was observing that I had one `music_albums` entry per track on each
compilation album after the album was scanned the second time (first time it
went in correctly).

The issue was that on reloading from the DB the field was not being initialised
so remained as `-1` when the entry came to be written back, which because the
field in the DB is `unsigned` ended up being stored as `0`, so when subsequent
lookups try to find the album it failed every time (since the 0 in the database
matches neither -1 nor the >0 correct value) and a fresh one is inserted for
every track.

Fix this by adding and using `{get,set}CompilationArtistId` corresponding to
the uses of `{get,set}ArtistId`. I broke out `getCompilationArtistId` from the
within exiting `getArtistId` implementation.