Transport Editor for DVB-C
The transport editor of mythtv-setup can now edit the
modulation system of a DVB-C transport stream multiplex
to any of the three supported values.
The list of choices and the default value for the symbol rate
are now correct for DVB-C and the same as presented in the
"Full Scan (Tuned)" pane of the scanner.