Duplicate channel number
In a channel scan, channels with duplicate channel numbers now
get a suffix A, B etc to make them unique.
In DVB the broadcasters can define logical channel numbers;
this are usuallly numbers from 1 upwards.
In ATSC a channel number can be a major/minor number, e.g. 7_2.
An ATSC channel number can also be defined by the broadcaster
as a single number from 1 upwards, similar to DVB.
If there is nothing defined then the service ID of the channel
is used as channel number. A service ID is generally unique
within a network but especially on satellites it can happen that
thereare different channels with the same service id.
When receiving OTA signals via aerials/antenna's it can happen
that the same multiplex is received from different transmitters.
This will also lead to duplicate channel numbers.
It is now easier to recognize duplicate channels because they
are shown next to each other in the program guide and in
the mythtv-setup channel editor.

Refs #13472

(cherry picked from commit 405e049c761135fd27c9def46137809144e6f2e9)
Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>