Signal strength of scanned transports
Show the signal strength of the scanned transports in the transport list.
The transport list is shown if the "-v chanscan" option is given when running mythtv-setup.
The signal strengths are useful to determine which transport to choose when identical
transports can be received from different transmitters on different frequencies.

Refs #13472