Updated "Remove duplicates" channel scan option
Renamed the feature from "Remove duplicate channels" to "Remove duplicates".
Changed the default for this option to Selected/Checked.
Removed the check on individual channels across all scanned channels.
The implementation does not check for original network ID plus transport ID
on a per-transport basis, as suggested in ticket #12107 for DVB, but it checks
this on all channels in the transport. The implementation is also expected to work for ATSC.
Thanks to John Pilkington for numerous tests in the daily changing UK Freeview landscape.

Refs #13472
Fixes #12107