Use const_iterator for QMap m_encoderList
Use const_iterator, constBegin, constEnd and constFind while accessing
the tvList and pointers to the tvList such as m_encoderList and m_tvList.
This fixes the problem of crashing while deleting a recording.
When deleting, the m_encoderList is accessed by more than one thread.
The QMap is documented to be re-entrant but this appears to be only the case
when the QMap is accessed with the const variants of the member functions.

Fixes #13571