Do not use exsting multiplex tuning data in a Full Scan
In a Full Scan the list of frequencies to be scanned and
the tuning parameters are generated from tables.
Previously, the tuning parameters were overwritten by the values
in the database if there was already a multiplex with that
frequency present in the database.
The problem with this was that if the information in the database
was not correct then that frequency could never be tuned again
unless the multiplex was deleted with the Transport Editor or
when a new video source, without multiplexes,  was created.