Simplify channel numbering
Simplify the channel numbering scheme to the following:
- Use the Logical Channel Number
- If that is not present then:
  - For ATSC, use the Major-Minor channel number
  - For DVB and everything else, use the service ID
If this results in a duplicate, a channel number that is
already present in the video source, then the channel number
gets a suffix A, B, or C etc to make it unique.
The advantage of this is that a channel has a more
predictable location in the program guide.

Refs #13472