Channelscan log output written with LOG
All mythtv-setup channel scanning log and debug output is now
written with the LOG utility. This is now possible because
the limit on the length of a log message, previously 2kbyte,
has been removed.
Previously, the channel scanner used both LOG messages and
direct writes to standard output with the iostream cout.
This caused irregular mixing of both output streams and
errors in the message ordering.