Consolidate formatting of time strings.
This started as a task to eliminate the obsolete QString::sprintf
function, and evolved into a task to consolidate the formatting of
tine strings through out the code.  There are now two functions to
call to get a formatted time.  The only difference is whether you pass
a time in seconds or in milliseconds.  The formatting argument to
these functions is the same as for the QTime::toString formatting
argument.