Fix "Full Scan" for DVB-T only tuners
The "Full Scan" for tuners that can do only DVB-T and not DVB-T2
was done correct but the modulation system was not entered in
the transport in the database. This is now fixed.
This gave problems in mythbackend when used with DVB-T/T2 tuners
because these need the modulation system information.

(cherry picked from commit a618b675fdd3b388221159e4d72cf15e56b2dfe4)
Signed-off-by: Klaas de Waal <kdewaal@mythtv.org>