Fix "Full Scan" for DVB-T only tuners
The "Full Scan" for tuners that can do only DVB-T and not DVB-T2
was done correct but the modulation system was not entered in
the transport in the database. This is now fixed.
This gave problems in mythbackend when used with DVB-T/T2 tuners
because these need the modulation system information.