UK EIT fixup fix for missing description
The UK EIT fixup did remove the description completely if there
was a year present anywhere in the subtitle or description.
If there was a year present in the title then the description was
replaced by the right part of the title starting with the year,
thereby also discarding the existing description.
This code is now deactivated and will be deleted when the correctness
of this fix is proven and there are no regressions.

Refs #13607